requirements.txt
# Build tools (install first)
setuptools>=70.0.0
wheel>=0.45.0

# Web framework
fastapi>=0.128.2
uvicorn[standard]>=0.40.0
python-multipart>=0.0.22

# Image processing
pillow>=12.1.0
numpy>=1.26.4,<2.0.0  # NumPy 2.x requires Python 3.11+, using 1.x for Python 3.10+ compatibility

# Deep learning
torch>=2.7.0
torchvision>=0.22.0

# Medical imaging
nibabel>=5.3.2

# Utilities
pydantic>=2.12.5
aiofiles>=25.1.0
scipy>=1.15.0,<1.17.0  # SciPy 1.17.0+ requires Python 3.11+, using 1.15.x for Python 3.10+ compatibility
